removed the previous initial timestep setting from the first iteration in the while loop and changed t to size_t.

+++ src/sph_solver.h
@@ -17,7 +17,7 @@ class SphSolver {
   // Time related variables
   bool adaptiveTimestepBool = false;
 
-  int t = 0;
+  size_t t = 0;
 
   int outputFrequency;
